+++ src/templates/imports.json
@@ -1,27 +1,21 @@
 {
-  "@floating-ui/core": "https://ga.jspm.io/npm:@floating-ui/core@0.6.2/dist/floating-ui.core.esm.development.js",
-  "@floating-ui/dom": "https://ga.jspm.io/npm:@floating-ui/dom@0.4.5/dist/floating-ui.dom.esm.development.js",
-  "@floating-ui/react-dom": "https://ga.jspm.io/npm:@floating-ui/react-dom@0.6.3/dist/floating-ui.react-dom.esm.js",
-  "@floating-ui/react-dom-interactions": "https://ga.jspm.io/npm:@floating-ui/react-dom-interactions@0.3.1/dist/floating-ui.react-dom-interactions.esm.js",
-  "aria-hidden": "https://ga.jspm.io/npm:aria-hidden@1.2.3/dist/es2015/index.js",
-  "htm": "https://ga.jspm.io/npm:htm@3.1.1/dist/htm.module.js",
-  "htm/react": "https://ga.jspm.io/npm:htm@3.1.1/react/index.module.js",
-  "point-in-polygon": "https://ga.jspm.io/npm:point-in-polygon@1.1.0/index.js",
-  "react-merge-refs": "https://ga.jspm.io/npm:react-merge-refs@1.1.0/dist/react-merge-refs.esm.js",
-  "click-to-react-component": "https://ga.jspm.io/npm:click-to-react-component@1.0.8/src/index.js",
-  "react-zoom-pan-pinch": "https://ga.jspm.io/npm:react-zoom-pan-pinch@3.0.6/dist/index.esm.js",
   "react": "https://ga.jspm.io/npm:react@17.0.2/index.js",
   "react-dom": "https://ga.jspm.io/npm:react-dom@17.0.2/index.js",
   "react-native": "https://unpkg.com/react-native-web-temp@0.17.6/dist/index.js",
   "react-native-web": "https://unpkg.com/react-native-web-temp@0.17.6/dist/index.js",
   "react-native-svg": "https://ga.jspm.io/npm:react-native-svg-web@1.0.9/index.js",
-  "create-react-class": "https://ga.jspm.io/npm:create-react-class@15.7.0/index.js",
-  "css-in-js-utils/lib/hyphenateProperty": "https://ga.jspm.io/npm:css-in-js-utils@2.0.1/lib/hyphenateProperty.js",
-  "css-in-js-utils/lib/isPrefixedValue": "https://ga.jspm.io/npm:css-in-js-utils@2.0.1/lib/isPrefixedValue.js",
-  "fbjs/lib/ExecutionEnvironment": "https://ga.jspm.io/npm:fbjs@3.0.2/lib/ExecutionEnvironment.js",
-  "fbjs/lib/invariant": "https://ga.jspm.io/npm:fbjs@3.0.2/lib/invariant.js",
-  "fbjs/lib/warning": "https://ga.jspm.io/npm:fbjs@3.0.2/lib/warning.js",
+  "react-zoom-pan-pinch": "https://ga.jspm.io/npm:react-zoom-pan-pinch@3.0.6/dist/index.esm.js",
   "hyphenate-style-name": "https://ga.jspm.io/npm:hyphenate-style-name@1.0.4/index.js",
+  "normalize-css-color": "https://ga.jspm.io/npm:normalize-css-color@1.0.2/index.js",
+  "create-react-class": "https://ga.jspm.io/npm:create-react-class@15.7.0/index.js",
+  "object-assign": "https://ga.jspm.io/npm:object-assign@4.1.1/index.js",
+  "prop-types": "https://ga.jspm.io/npm:prop-types@15.8.0/index.js",
+  "scheduler": "https://ga.jspm.io/npm:scheduler@0.20.2/index.js",
+  "fbjs/lib/warning": "https://ga.jspm.io/npm:fbjs@3.0.2/lib/warning.js",
+  "fbjs/lib/invariant": "https://ga.jspm.io/npm:fbjs@3.0.2/lib/invariant.js",
+  "fbjs/lib/ExecutionEnvironment": "https://ga.jspm.io/npm:fbjs@3.0.2/lib/ExecutionEnvironment.js",
+  "css-in-js-utils/lib/isPrefixedValue": "https://ga.jspm.io/npm:css-in-js-utils@2.0.1/lib/isPrefixedValue.js",
+  "css-in-js-utils/lib/hyphenateProperty": "https://ga.jspm.io/npm:css-in-js-utils@2.0.1/lib/hyphenateProperty.js",
   "inline-style-prefixer/lib/createPrefixer": "https://unpkg.com/inline-style-prefixer@6.0.1/es/createPrefixer.js",
   "inline-style-prefixer/lib/plugins/backgroundClip": "https://unpkg.com/inline-style-prefixer@6.0.1/es/plugins/backgroundClip.js",
   "inline-style-prefixer/lib/plugins/crossFade": "https://unpkg.com/inline-style-prefixer@6.0.1/es/plugins/crossFade.js",
@@ -36,9 +30,5 @@
   "inline-style-prefixer/lib/plugins/logical": "https://unpkg.com/inline-style-prefixer@6.0.1/es/plugins/logical.js",
   "inline-style-prefixer/lib/plugins/position": "https://unpkg.com/inline-style-prefixer@6.0.1/es/plugins/position.js",
   "inline-style-prefixer/lib/plugins/sizing": "https://unpkg.com/inline-style-prefixer@6.0.1/es/plugins/sizing.js",
-  "inline-style-prefixer/lib/plugins/transition": "https://unpkg.com/inline-style-prefixer@6.0.1/es/plugins/transition.js",
-  "normalize-css-color": "https://ga.jspm.io/npm:normalize-css-color@1.0.2/index.js",
-  "object-assign": "https://ga.jspm.io/npm:object-assign@4.1.1/index.js",
-  "prop-types": "https://ga.jspm.io/npm:prop-types@15.8.0/index.js",
-  "scheduler": "https://ga.jspm.io/npm:scheduler@0.20.2/index.js"
+  "inline-style-prefixer/lib/plugins/transition": "https://unpkg.com/inline-style-prefixer@6.0.1/es/plugins/transition.js"
 }
